+++ src/main/java/org/newdawn/slick/geom/Ellipse.java
@@ -164,7 +164,7 @@ public class Ellipse extends Shape {
         }
         points = new float[tempPoints.size()];
         for(int i=0;i<points.length;i++) {
-            points[i] = ((Float)tempPoints.get(i)).floatValue();
+            points[i] = tempPoints.get(i).floatValue();
         }
     }
