+++ src/main/java/org/newdawn/slick/geom/Vector2f.java
@@ -372,7 +372,7 @@ public strictfp class Vector2f implements Serializable {
         float dx = other.getX() - getX();
         float dy = other.getY() - getY();
 
-        return (float) (dx*dx)+(dy*dy);
+        return dx*dx +(dy*dy);
     }
 
     /**
